requirements.txt
flask
python-dotenv
google-cloud-firestore
google-cloud-storage
google-generativeai
firebase-admin
